refactor: use `__fixtures__` for code used only for jest tests internally
there are some mocks that we use internally that we don't need to export out.
