+++ packages/notification-services-controller/src/NotificationServicesController/services/feature-announcements.test.ts
@@ -3,7 +3,7 @@ import {
   getFeatureAnnouncementUrl,
 } from './feature-announcements';
 import { createMockFeatureAnnouncementAPIResult } from '../mocks/mock-feature-announcements';
-import { mockFetchFeatureAnnouncementNotifications } from '../mocks/mockServices';
+import { mockFetchFeatureAnnouncementNotifications } from '../__fixtures__/mockServices';
 import { TRIGGER_TYPES } from '../constants/notification-schema';
 
 // Mocked type for testing, allows overwriting TS to test erroneous values
