+++ packages/notification-services-controller/src/NotificationServicesPushController/__fixtures__/mockServices.ts
@@ -1,6 +1,6 @@
 import nock from 'nock';
 
-import { getMockUpdatePushNotificationLinksResponse } from './mockResponse';
+import { getMockUpdatePushNotificationLinksResponse } from '../mocks/mockResponse';
 
 type MockReply = {
   status: nock.StatusCode;
