+++ packages/notification-services-controller/src/NotificationServicesPushController/web/push-utils.test.ts
@@ -11,7 +11,7 @@ import {
 import * as PushWebModule from './push-utils';
 import { processNotification } from '../../NotificationServicesController';
 import { createMockNotificationEthSent } from '../../NotificationServicesController/mocks/mock-raw-notifications';
-import { buildPushPlatformNotificationsControllerMessenger } from '../mocks/mockMessenger';
+import { buildPushPlatformNotificationsControllerMessenger } from '../__fixtures__/mockMessenger';
 
 jest.mock('firebase/app');
 jest.mock('firebase/messaging');
